Ultimos Ajustes

+++ Controllers/roleController.js
@@ -7,18 +7,20 @@ class RoleController {
             const result = await Role.findAll({
                 attributes: ["id", "name"]
             });
+            if (result.length == 0) {throw new Error("No hay roles!")}
             res.status(200).send({ success: true, message: result });
         } catch (error) {
-            res.status(400).send({ success: false, message: error })
+            res.status(400).send({ success: false, message: error.message })
         }
     }
     async getRoleById(req, res) {
         try {
             const { id } = req.params;
             const result = await Role.findByPk(id);
+            if (result == null) {throw new Error("No se encontró el rol!")}
             res.status(200).send({ success: true, message: result })
         } catch (error) {
-            res.status(400).send({ success: false, message: error });
+            res.status(400).send({ success: false, message: error.message });
         }
     }
     async createRole(req, res) {
@@ -36,6 +38,7 @@ class RoleController {
         try {
             const { id } = req.params;
             const { name } = req.body;
+            if (!await Role.findByPk(id)) {throw new Error("No se encontró el rol!")}
             const result = await Role.update(
                 { name },
                 {
@@ -46,12 +49,13 @@ class RoleController {
             );
             res.status(200).send({ success: true, message: "rol modificado con exito" })
         } catch (error) {
-            res.status(400).send({ success: false, message: error })
+            res.status(400).send({ success: false, message: error.message })
         }
     }
     async deleteRole(req, res) {
         try {
             const { id } = req.params
+            if (!await Role.findByPk(id)) {throw new Error("No se encontró el rol!")}
             const result = await Role.destroy({
                 where: {
                     id
@@ -59,7 +63,7 @@ class RoleController {
             })
             res.status(200).send({success: true, message: "rol eliminado con exito"})
         } catch (error) {
-            res.status(400).send({ success: false, message: error })
+            res.status(400).send({ success: false, message: error.message })
         }
     }
 }
